+++ packages/ragbits-chat/src/ragbits/chat/interface/types.py
@@ -104,6 +104,12 @@ class ChatResponse(BaseModel):
         """
         return cast(StateUpdate, self.content) if self.type == ChatResponseType.STATE_UPDATE else None
 
+    def as_conversation_id(self) -> str | None:
+        """
+        Return the content as ConversationID if this is a conversation id, else None.
+        """
+        return cast(str, self.content) if self.type == ChatResponseType.CONVERSATION_ID else None
+
     def as_live_update(self) -> LiveUpdate | None:
         """
         Return the content as LiveUpdate if this is a live update, else None.