+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added synchronous and asynchronous Python api client
 - Loading indicator, delayed message buttons, integrated abort controller (#641)
 - Added Eslint, Prettier & CI/CD for Ragbits API Clients (#604)
 - Excluded API connection logic to 2 modules: ragbits-api-client and ragbits-api-client-react which implements hooks for ragbits-api-client (#582)